+++ src/lib/app-provider/queries/simulateFee.ts
@@ -1,3 +1,4 @@
+import type { Coin } from "@cosmjs/amino";
 import { SigningCosmWasmClient } from "@cosmjs/cosmwasm-stargate";
 import { useWallet } from "@cosmos-kit/react";
 import { useQuery } from "@tanstack/react-query";
@@ -11,7 +12,7 @@ import type {
   HumanAddr,
   Option,
 } from "lib/types";
-import { composeStoreCodeMsg } from "lib/utils";
+import { composeStoreCodeMsg, composeStoreCodeProposalMsg } from "lib/utils";
 
 interface SimulateQueryParams {
   enabled: boolean;
@@ -94,7 +95,7 @@ export const useSimulateFeeForStoreCode = ({
     const client = await getCosmWasmClient();
     if (!client) throw new Error("Fail to get client");
 
-    const submitStoreCodeProposalMsg = async () => {
+    const submitStoreCodeMsg = async () => {
       return composeStoreCodeMsg({
         sender: address as HumanAddr,
         wasmByteCode: new Uint8Array(await wasmFile.arrayBuffer()),
@@ -102,7 +103,7 @@ export const useSimulateFeeForStoreCode = ({
         addresses,
       });
     };
-    const craftMsg = await submitStoreCodeProposalMsg();
+    const craftMsg = await submitStoreCodeMsg();
     return (await client.simulate(address, [craftMsg], undefined)) as Gas;
   };
   return useQuery({
@@ -112,6 +113,94 @@ export const useSimulateFeeForStoreCode = ({
       wasmFile,
       permission,
       addresses,
+    ],
+    queryFn: async () => simulateFn(),
+    enabled,
+    retry: false,
+    refetchOnReconnect: false,
+    refetchOnWindowFocus: false,
+    onSuccess,
+    onError,
+  });
+};
+
+interface SimulateQueryParamsForProposalStoreCode {
+  enabled: boolean;
+  title: string;
+  description: string;
+  runAs: Addr;
+  initialDeposit: Coin;
+  unpinCode: boolean;
+  builder: string;
+  source: string;
+  codeHash: string;
+  wasmFile: Option<File>;
+  permission: AccessType;
+  addresses: Addr[];
+  onSuccess?: (gas: Gas<number> | undefined) => void;
+  onError?: (err: Error) => void;
+}
+
+export const useSimulateFeeForProposalStoreCode = ({
+  enabled,
+  title,
+  description,
+  runAs,
+  initialDeposit,
+  unpinCode,
+  builder,
+  source,
+  codeHash,
+  wasmFile,
+  permission,
+  addresses,
+  onSuccess,
+  onError,
+}: SimulateQueryParamsForProposalStoreCode) => {
+  const { address, getCosmWasmClient, currentChainName } = useWallet();
+
+  const simulateFn = async () => {
+    if (!address) throw new Error("Please check your wallet connection.");
+    if (!wasmFile) throw new Error("Fail to get Wasm file");
+
+    const client = await getCosmWasmClient();
+    if (!client) throw new Error("Fail to get client");
+
+    const submitStoreCodeProposalMsg = async () => {
+      return composeStoreCodeProposalMsg({
+        proposer: address as HumanAddr,
+        title,
+        description,
+        runAs: runAs as Addr,
+        wasmByteCode: new Uint8Array(await wasmFile.arrayBuffer()),
+        permission,
+        addresses,
+        unpinCode,
+        source,
+        builder,
+        codeHash: Uint8Array.from(Buffer.from(codeHash, "hex")),
+        initialDeposit,
+      });
+    };
+
+    const craftMsg = await submitStoreCodeProposalMsg();
+    return (await client.simulate(address, [craftMsg], undefined)) as Gas;
+  };
+
+  return useQuery({
+    queryKey: [
+      "simulate_fee_store_code_proposal",
+      currentChainName,
+      runAs,
+      initialDeposit,
+      unpinCode,
+      builder,
+      source,
+      codeHash,
+      wasmFile,
+      permission,
+      addresses,
+      enabled,
     ],
     queryFn: async () => simulateFn(),
     enabled,