+++ src/lib/app-provider/types.ts
@@ -1,3 +1,4 @@
 export interface AppConstants {
   gasAdjustment: number;
+  maxGasLimit: number;
 }