+++ src/lib/styles/theme/components/checkbox.ts
@@ -1,0 +1,17 @@
+import type { ComponentStyleConfig } from "@chakra-ui/react";
+
+const violet = "violet.light";
+export const Checkbox: ComponentStyleConfig = {
+  baseStyle: {
+    control: {
+      _checked: {
+        bg: violet,
+        borderColor: violet,
+        _hover: {
+          bg: violet,
+          borderColor: violet,
+        },
+      },
+    },
+  },
+};