+++ src/lib/utils/tx/composeMsg.ts
@@ -1,5 +1,6 @@
 import type { Coin } from "@cosmjs/stargate";
 import { ParameterChangeProposal } from "cosmjs-types/cosmos/params/v1beta1/params";
+import { StoreCodeProposal } from "cosmjs-types/cosmwasm/wasm/v1/proposal";
 
 import { microfy } from "../formatter";
 import { typeUrlDict } from "lib/data";
@@ -9,6 +10,7 @@ import type {
   AccessType,
   Addr,
   Token,
+  HumanAddr,
 } from "lib/types";
 import { MsgType } from "lib/types";
 
@@ -39,6 +41,7 @@ export const composeStoreCodeMsg = ({
     instantiatePermission: {
       permission,
       addresses,
+      address: "" as Addr,
     },
   });
 
@@ -76,9 +79,69 @@ export const composeSubmitWhitelistProposalMsg = ({
     },
     initialDeposit: [
       {
-        ...initialDeposit,
+        denom: initialDeposit.denom,
         amount: microfy((initialDeposit.amount || 0) as Token).toFixed(0),
       },
     ],
     proposer,
   });
+
+interface StoreCodeProposalMsgArgs {
+  proposer: HumanAddr;
+  title: string;
+  description: string;
+  runAs: Addr;
+  wasmByteCode: Uint8Array;
+  permission: AccessType;
+  addresses: Addr[];
+  unpinCode: boolean;
+  source: string;
+  builder: string;
+  codeHash: Uint8Array;
+  initialDeposit: Coin;
+}
+
+export const composeStoreCodeProposalMsg = ({
+  proposer,
+  title,
+  description,
+  runAs,
+  wasmByteCode,
+  permission,
+  addresses,
+  unpinCode,
+  source,
+  builder,
+  codeHash,
+  initialDeposit,
+}: StoreCodeProposalMsgArgs): ComposedMsg => {
+  return composeMsg(MsgType.SUBMIT_PROPOSAL, {
+    content: {
+      typeUrl: "/cosmwasm.wasm.v1.StoreCodeProposal",
+      value: Uint8Array.from(
+        StoreCodeProposal.encode({
+          title: title.trim(),
+          description: description.trim(),
+          runAs,
+          wasmByteCode,
+          instantiatePermission: {
+            permission,
+            addresses,
+            address: "" as Addr,
+          },
+          unpinCode,
+          source,
+          builder,
+          codeHash,
+        }).finish()
+      ),
+    },
+    initialDeposit: [
+      {
+        denom: initialDeposit.denom,
+        amount: microfy((initialDeposit.amount || 0) as Token).toFixed(0),
+      },
+    ],
+    proposer,
+  });
+};