+++ src/pages/[network]/proposal/store-code.tsx
@@ -1,0 +1,3 @@
+import StoreCodeProposal from "pages/proposal/store-code";
+
+export default StoreCodeProposal;