+++ src/pages/proposal/store-code.tsx
@@ -1,0 +1,3 @@
+import StoreCodePropolsal from "lib/pages/proposal/store-code";
+
+export default StoreCodePropolsal;